evotraders/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "evotraders"
version = "0.1.0"
description = "EvoTraders: A self-evolving multi-agent trading system"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "EvoTraders Team <dengjiaji.djj@alibaba-inc.com>"}
]
keywords = ["trading", "ai", "multi-agent", "fintech", "algorithmic-trading"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "agentscope>=1.0.8",
    "reme-ai>=0.2.0.3",
    "asyncio>=3.4.3",
    "rich>=13.6.0",
    "websockets>=12.0",
    "websocket-client>=1.6.0",
    "python-dotenv>=1.0.0",
    "finnhub-python>=2.4.25",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pandas-market-calendars>=5.0.0",
    "typer>=0.12.5",
    "openai>=2.9.0",
    "anthropic>=0.20.0",
    "dotenv",
    "typer",
]


[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "ruff>=0.6.9",
    "black>=25.0.0"
]

[project.urls]
Homepage = "http://trading.evoagents.cn"
Repository = "https://github.com/agentscope-ai/agentscope-samples/evotraders"
Documentation = "https://github.com/agentscope-ai/agentscope-samples/evotraders/README.md"
"Bug Tracker" = "https://github.com/agentscope-ai/agentscope-samples/issues"

[project.scripts]
evotraders = "backend.cli:app"

[tool.setuptools]
packages = ["backend", "backend.agents", "backend.config",
            "backend.data", "backend.llm",
            "backend.tools", "backend.utils", "backend.services"]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

